requirements.txt
streamlit==1.41.1
google-generativeai==0.8.3
python-dotenv==1.0.1
langchain==0.3.15
langchain-community==0.3.15
chromadb==0.5.23
faiss-cpu==1.9.0.post1
langchain_google_genai==2.0.7
gunicorn==23.0.0
langchain_text_splitters==0.3.5
langchain_chroma==0.1.4
deep-translator==1.11.4
openpyxl==3.1.2